+++ tests/e2e/get_query_test.go
@@ -1,0 +1,43 @@
+package tests_e2e
+
+import (
+	"fmt"
+	"strings"
+	"testing"
+
+	. "github.com/dhuan/mock/tests/e2e/utils"
+)
+
+func Test_E2E_GetQuery(t *testing.T) {
+	RunTestWithNoConfigAndWithArgs(
+		t,
+		[]string{
+			"--route foo/bar",
+			fmt.Sprintf("--exec '%s'", strings.Join([]string{
+				`{{MOCK_EXECUTABLE}} get-query someKey | {{MOCK_EXECUTABLE}} write`,
+			}, ";")),
+		},
+		"GET",
+		"foo/bar?someKey=someValue",
+		nil,
+		nil,
+		StringMatches("someValue"),
+	)
+}
+
+func Test_E2E_GetQuery_ExitCode1WhenKeyDoesNotExist(t *testing.T) {
+	RunTestWithNoConfigAndWithArgs(
+		t,
+		[]string{
+			"--route foo/bar",
+			fmt.Sprintf("--exec '%s'", strings.Join([]string{
+				`{{MOCK_EXECUTABLE}} get-query someKey ; printf $? | mock write`,
+			}, ";")),
+		},
+		"GET",
+		"foo/bar?foo=bar",
+		nil,
+		nil,
+		StringMatches("1"),
+	)
+}